リソース追加、3d描画実装

+++ Program/Modules/Core/Platform/Win/PlatformImpl.h
@@ -26,10 +26,10 @@ CANDY_NAMESPACE_BEGIN
 using vector_type = __m128;
 using vectori_type = __m128i;
 
-#define CANDY_MAKE_FOURCC(x, y, z, w) (((w) << 24) | ((z) << 16) | ((y) << 8) | (x))
-#define CANDY_MAKE_FOURCC_STR(str) (((str)[3] << 24) | ((str)[2] << 16) | ((str)[1] << 8) | (str)[0])
+#define CANDY_MAKE_FOURCC(x, y, z, w) (((u32)(w) << 24) | ((u32)(z) << 16) | ((u32)(y) << 8) | (u32)(x))
+#define CANDY_MAKE_FOURCC_STR(str) (((u32)(str)[3] << 24) | ((u32)(str)[2] << 16) | ((u32)(str)[1] << 8) | (u32)(str)[0])
 
-#define CANDY_MM_SHUFFLE(fp0, fp1, fp2, fp3) (((fp3) << 6) | ((fp2) << 4) | ((fp1) << 2) | ((fp0)))
+#define CANDY_MM_SHUFFLE(fp0, fp1, fp2, fp3) (((u32)(fp3) << 6) | ((u32)(fp2) << 4) | ((u32)(fp1) << 2) | ((u32)(fp0)))
 
 CANDY_NAMESPACE_END
